% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvefinder.R
\name{CurveFinder}
\alias{CurveFinder}
\title{CurveFinder: Estimate curve and morphologically
relevant coordinates in spatial transcriptomics data}
\usage{
CurveFinder(xy, knn = 5, prune.outlier = NULL, loop = FALSE)
}
\arguments{
\item{normalized_counts}{A matrix containing
normalized data with genes on rows and cells
on columns}

\item{meta.data}{A data frame containing meta data for each cell.}

\item{fixed.effects}{The columns in meta.data corresponding to the fixed effects. In
a typical case, this would be the condition of interest.}

\item{random.effects}{The columns in meta.data corresponding to the random effects.
In a typical use case this would be the column containing patient ID.}

\item{clusters}{The column containing the cell-type annotation.}

\item{d}{The number of PCs to use.}

\item{truncate}{Whether or not to round negative distances to 0.}

\item{min.count.per.cell}{The minimum number of cells per cluster to perform the estimation.}

\item{weights}{An optional vector of length equal to the number of genes specifying the weight
to place on each gene in the distance estimate.}
}
\value{
A list with components
\itemize{
\item \code{results} - A data frame containing the cell
type, estimated distance, and other statistics such as p-value.
\item \code{vals} For each cell type a list of more detailed
information (such as raw data) and coefficients for each PC are
included.
}
}
\description{
An automa
}
