% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper1D.R
\name{get_mapper_graph}
\alias{get_mapper_graph}
\title{Mapper Algorithm}
\usage{
get_mapper_graph(
  data,
  filter_values,
  num_intervals = 10,
  percent_overlap = 30,
  eps = 0.1,
  min_pts = 50
)
}
\arguments{
\item{data}{A numeric matrix (n x 2) representing spatial coordinates.}

\item{filter_values}{A numeric vector (n) of the filter values on the data.}

\item{num_intervals}{An integer specifying the resolution of the Mapper
algorithm (how many level sets to cluster)}

\item{percent_overlap}{A numeric between 0 and 100 specifying the gain of
Mapper algorithm (what percent should level sets overlap)}

\item{eps}{A numeric specifying the epsilon radius that defines neighboring
points in DBSCAN.}

\item{min_pts}{An integer specifying the minimum cluster size for DBSCAN.}
}
\value{
igraph object containing the Mapper graph
}
\description{
See the following article for more details:
Singh, Gurjeet, Facundo Mémoli, and Gunnar E. Carlsson. "Topological methods
for the analysis of high dimensional data sets and 3d object recognition."
SPBG. 2007.
}
\examples{
my_data <- makeSwissRoll()
scaled_data <- scale(my_data)
pc <- prcomp(scaled_data, center = FALSE, scale = FALSE)
mapper_graph <- get_mapper_graph(scaled_data, pc$x[, 1], num_intervals = 7,
                                 eps = 0.2, min_pts = 5)
get_mapper_plot(scaled_data, mapper_graph)
}
