% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_detection.R
\name{is_loop}
\alias{is_loop}
\title{Determines if a k-nearest neighbors graph of a dataset forms a loop.}
\usage{
is_loop(xy, knn_graph)
}
\arguments{
\item{xy}{A numeric matrix or data frame with two columns representing the x
and y coordinates of the data points.}

\item{knn_graph}{A k-nearest neighbors graph of `xy`.}
}
\value{
Whether the k-nearest neighbors graph is a loop.
}
\description{
Loops are detected as large cycles in the output of the Mapper algorithm.
}
\examples{
my_data <- makeSwissRoll()
my_graph <- dimRed:::makeKNNgraph(x = my_data, k = 5, eps = 0)
testthat::expect_false(is_loop(my_data, my_graph))
}
