% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_detection.R
\name{get_mapper_graph}
\alias{get_mapper_graph}
\title{Mapper Algorithm}
\usage{
get_mapper_graph(
  graph,
  filter_values,
  num_intervals = 10,
  percent_overlap = 33
)
}
\arguments{
\item{graph}{An igraph object (n vertices).}

\item{filter_values}{A numeric vector (n) of filter values, specifying how to
order the vertices of the graph and thus group them into level sets.}

\item{num_intervals}{An integer specifying the resolution of the Mapper
algorithm (the number of level sets).}

\item{percent_overlap}{A numeric between 0 and 100 specifying the gain of
Mapper algorithm (the percentage that level sets should overlap).}
}
\value{
igraph object containing the Mapper graph
}
\description{
See the following article for more details:
Singh, Gurjeet, Facundo Mémoli, and Gunnar E. Carlsson. "Topological methods
for the analysis of high dimensional data sets and 3d object recognition."
SPBG. 2007.
}
\examples{
my_data <- makeSwissRoll()
my_graph <- dimRed:::makeKNNgraph(x = my_data, k = 5, eps = 0)
mapper_graph <- get_mapper_graph(my_graph, my_data[, 1])
get_mapper_plot(my_data, mapper_graph)
}
