% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphoGAM.R, R/zzz.R
\docType{package}
\name{MorphoGAM}
\alias{MorphoGAM-package}
\alias{MorphoGAM}
\title{MorphoGAM}
\usage{
MorphoGAM(
  Y,
  curve.fit,
  design,
  shrinkage = TRUE,
  min.count.per.gene = 10,
  return.fx = TRUE,
  offset = NULL
)
}
\description{
DESCRIPTION
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
